refactor(builder): relocate identity cells to team-builder/shared
Pure file moves with import updates - no logic changes. Moves the
identity cell components and helpers out of lanes/identity/cells/ into
a flat shared/ namespace so they can be consumed by upcoming
compact-row and grid-row layout variants without crossing lane folders.

- 12 files moved via git mv (history preserved)
- 5 lane files updated for new import paths
- shared/fields/ subfolder for the 5 loadout field components
- All exported symbol names unchanged
- 1966 tests pass, typecheck clean
