+++ apps/web/src/components/team-builder/lanes/identity/identity-mid-stack.tsx
@@ -4,11 +4,11 @@ import { useState } from "react";
 
 import { FieldErrors } from "../../validation/field-error";
 import { SpeciesPickerDialog } from "../../pickers/species-picker-dialog";
-import { useIdentityState } from "./use-identity-state";
-import { FormCells } from "./cells/form-cells";
-import { MetaBar } from "./cells/meta-bar";
-import { SpriteSection } from "./cells/sprite-section";
-import { type IdentityLayoutProps } from "./identity-layout-props";
+import { useIdentityState } from "../../shared/use-identity-state";
+import { FormCells } from "../../shared/fields";
+import { MetaBar } from "../../shared/meta-bar";
+import { SpriteSection } from "../../shared/sprite-section";
+import { type IdentityLayoutProps } from "../../shared/identity-layout-props";
 
 // =============================================================================
 // IdentityMidStack — MidStack layout (slot < 1240px)
