+++ apps/web/src/components/team-builder/lanes/identity/identity-single-row.tsx
@@ -6,11 +6,11 @@ import { useCalcEnabled } from "../../calc/calc-state-context";
 import { cn } from "@/lib/utils";
 import { FieldErrors } from "../../validation/field-error";
 import { SpeciesPickerDialog } from "../../pickers/species-picker-dialog";
-import { useIdentityState } from "./use-identity-state";
-import { FormCells } from "./cells/form-cells";
-import { MetaBar } from "./cells/meta-bar";
-import { SpriteSection } from "./cells/sprite-section";
-import { type IdentityLayoutProps } from "./identity-layout-props";
+import { useIdentityState } from "../../shared/use-identity-state";
+import { FormCells } from "../../shared/fields";
+import { MetaBar } from "../../shared/meta-bar";
+import { SpriteSection } from "../../shared/sprite-section";
+import { type IdentityLayoutProps } from "../../shared/identity-layout-props";
 
 // =============================================================================
 // IdentitySingleRow — compact layout (slot ≥ 1240px)
